refactor: move manufacturing validation fcns to Machine class refactor: move sentry enable/disable to api.manufacturing
